+++ examples/language_model/gpt-3/dygraph/run_pretrain.py
@@ -333,6 +333,7 @@ def do_train(args):
                 data_world_rank=data_world_rank,
                 max_seq_len=args.max_seq_len,
                 eos_id=tokenizer.eos_token_id,
+                old_version_accumulate_compatible=True,
             )
             # Bug fix, if not call valid_data_loader, the enumerate will call valid_data_loader
             # many times. and start a new random dataloader.
